backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.0.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-fernet-fields==0.6
django-filter==25.1
django-stubs==5.2.2
django-stubs-ext==5.2.2
djangorestframework==3.16.1
djangorestframework-stubs==3.16.1
djangorestframework_simplejwt==5.5.1
djfernet==0.8.1
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
PyJWT==2.10.1
pyotp==2.9.0
PyYAML==6.0.2
qrcode==8.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
sqlparse==0.5.3
tomli==2.2.1
types-PyYAML==6.0.12.20250809
types-requests==2.32.4.20250809
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
